+++ src/main/java/com/lowdragmc/mbd2/common/data/MBDRecipeConditions.java
@@ -23,7 +23,8 @@ public final class MBDRecipeConditions {
         MBDRegistries.RECIPE_CONDITIONS.register(MachineLevelCondition.INSTANCE.getType(), MachineLevelCondition.class);
         MBDRegistries.RECIPE_CONDITIONS.register(MachineNBTCondition.INSTANCE.getType(), MachineNBTCondition.class);
         MBDRegistries.RECIPE_CONDITIONS.register(BlockCondition.INSTANCE.getType(), BlockCondition.class);
-        MBDRegistries.RECIPE_CONDITIONS.register(DayLightCondition.INSTANCE.getType(), DayLightCondition.class);
+        MBDRegistries.RECIPE_CONDITIONS.register(DayTimeCondition.INSTANCE.getType(), DayTimeCondition.class);
+        MBDRegistries.RECIPE_CONDITIONS.register(LightCondition.INSTANCE.getType(), LightCondition.class);
         MBDRegistries.RECIPE_CONDITIONS.register(RedstoneSignalCondition.INSTANCE.getType(), RedstoneSignalCondition.class);
         if (MBD2.isCreateLoaded()) {
             MBDRegistries.RECIPE_CONDITIONS.register(CreateRotationCondition.INSTANCE.getType(), CreateRotationCondition.class);