+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## v1.0.30
 * Fixed Forge Additional Capabilities doesn't work
+* Added Light Condition for `sky light`, `block light` and `can see sky`